Clean up mixup of customtext2 and customtext3 fields (for fresh installations only)
In previous commits in this plugin, a mixup of the customtext2 and customtext3 fields was introduced and never cleaned up since then.
https://github.com/emeneo/moodle-enrol_apply/issues/89 dealt with the fact that a DB upgrade step was missing and came to a preliminary end with the conclusion that admins should fix existing data manually when upgrading the plugin.

However, the plugin is still broken for fresh installations as well. This is because the existing code still pre-fills customtext2 with notification settings, but reads the custom label from there. As a teacher adding an enrolment instance to a course, I see the 'Custom label' field filled with '$@ALL@$' which not really makes sense.

This patch tries to clean up this mixup so that at least fresh installations (and installations where the admin has cleaned up the data manually) work well again.